docs: storybook v7 upgrade
