+++ .storybook/baklava-theme.js
@@ -1,0 +1,14 @@
+import { create } from '@storybook/theming';
+
+export default create({
+  base: 'light',
+  brandTitle: 'Baklava Design System',
+  brandUrl: 'https://github.com/Trendyol/baklava',
+  brandImage: 'https://user-images.githubusercontent.com/127687/230370329-71295041-f5fc-4e12-8bf5-f37e98831efb.svg',
+  colorPrimary: '#F27A1A',
+  colorSecondary: '#273142',
+  fontBase: 'RubikVariable, sans-serif',
+  fontCode:
+    '"Operator Mono","Fira Code Retina","Fira Code","FiraCode-Retina","Andale Mono","Lucida Console",Consolas,Monaco,monospace',
+  textColor: '#273142',
+});
