+++ custom-elements-manifest.config.mjs
@@ -1,5 +1,5 @@
 export default {
-  globs: ['src/components/**/!(*.test).ts'],
+  globs: ['src/components/**/!(*.test|*.stories).ts'],
   exclude: ['src/**/*.css', 'src/components/icon/icon-list.ts'],
   outdir: 'dist/',
   dev: false,
