+++ docs/baklava-release-flow.stories.mdx
@@ -1,16 +1,6 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from '@storybook/blocks';
 
-<Meta
-  title="Documentation/Release Flow"
-  parameters={{
-    viewMode: 'docs',
-    previewTabs: {
-      canvas: {
-        hidden: true,
-      },
-    },
-  }}
-/>
+<Meta title="Documentation/Release Flow" />
 
 # Baklava Release Flow
