+++ docs/how-to-customize-a-components-style.stories.mdx
@@ -1,18 +1,8 @@
-import { Meta, Description } from '@storybook/addon-docs';
+import { Meta } from '@storybook/blocks';
 import cssCustomPropertiesStorybookImage from './images/css-custom-properties-storybook.png'
 import cssCustomPropertiesComponentImage from './images/css-custom-properties-component.png'
 
-<Meta
-  title="Documentation/Customizing Components Style"
-  parameters={{
-    viewMode: 'docs',
-    previewTabs: {
-      canvas: {
-        hidden: true,
-      },
-    },
-  }}
-/>
+<Meta title="Documentation/Customizing Components Style" />
 
 # How to customize a component's style?
 > You can look at our [discussion](https://github.com/Trendyol/baklava/discussions/164) to know more about our decision process.
@@ -31,6 +21,7 @@ with prefixes in every component. Therefore, should be no prefixes in local vari
 However, if we want to add some flexibility, customizability, we define variables with components prefix like `--bl-badge-bg-color`.
 
 #### Example
+
 Let's take a badge component as an example:
 ```css
 /* bl-badge.css */
