+++ docs/how-to-use-playground.stories.mdx
@@ -1,17 +1,7 @@
-import { Meta, Description } from '@storybook/addon-docs';
+import { Meta } from '@storybook/blocks';
 import playgroundLandingPage from './images/playground-landing-page.png'
 
-<Meta
-  title="Documentation/Contributing Baklava/How To Use Playground"
-  parameters={{
-    viewMode: 'docs',
-    previewTabs: {
-      canvas: {
-        hidden: true,
-      },
-    },
-  }}
-/>
+<Meta title="Documentation/Contributing Baklava/How To Use Playground" />
 
 # How to use playground?
