+++ docs/requirements.stories.mdx
@@ -1,16 +1,6 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from '@storybook/blocks';
 
-<Meta
-  title="Documentation/Requirements"
-  parameters={{
-    viewMode: 'docs',
-    previewTabs: {
-      canvas: {
-        hidden: true,
-      },
-    },
-  }}
-/>
+<Meta title="Documentation/Requirements" />
 
 # Requirements
 
@@ -39,10 +29,10 @@ The remaining 5% can be supported by compiling javascript in build time and addi
 ### Modern Browser Support
 | Browser |  Supports ES2019 & web components |
 |---------|------|
-| Chrome  |  >=73  |
-| Firefox |  >=12.1 |
-| Safari  |  >=63  |
-| Edge    |  >=79  |
+| Chrome  |  `>=73`  |
+| Firefox |  `>=12.1` |
+| Safari  |  `>=63`  |
+| Edge    |  `>=79`  |
 
 ### Legacy Browser Support
 Supporting older browsers requires some steps described by the Lit documentation.
@@ -55,12 +45,12 @@ For compiling and transforming, you can install Baklava as an npm package and tr
 
 | Browser |  Compile JS	| Compile JS & load polyfills |
 |---------|------|-----|
-| Chrome  |  67-79	  | <67 |
-| Firefox |  10-12	 | <10|
-| Safari  |  63-71  | <63|
+| Chrome  |  67-79	  | `<67` |
+| Firefox |  10-12	 | `<10`|
+| Safari  |  63-71  | `<63` |
 | Edge    | 79  | - |
-| Edge Classic   | - | <=18 |
-| Internet Explorer   | -  | 11 |
+| Edge Classic   | - | `<=18` |
+| Internet Explorer   | -  | `11` |
 
 ## Pollyfills
